Move util to utils and improve
